+++ src/app/Http/Controllers/Admin/ContactFormValueAdminController.php
@@ -1,6 +1,7 @@
 <?php
 namespace VCComponent\Laravel\ConfigContact\Http\Controllers\Admin;
 
+use Exception;
 use Illuminate\Http\Request;
 use VCComponent\Laravel\ConfigContact\Repositories\ContactFormValueRepository;
 use VCComponent\Laravel\ConfigContact\Transformers\ContactFormValueTransformer;
@@ -38,6 +39,10 @@ class ContactFormValueAdminController extends ApiController
 
     public function show(Request $request, $id)
     {
+        $contact_form_value = $this->contact_form_value_entity->find($id);
+        if (!$contact_form_value) {
+            throw new Exception('Contact form value not found');
+        }
         $perpage            = $request->has('per_page') ? $request->get('per_page') : 15;
         $contact_form_value = $this->contact_form_value_repository->where('contact_form_id', $id)->paginate($perpage);
         return $this->response->paginator($contact_form_value, $this->contact_form_value_transformer);
@@ -45,6 +50,10 @@ class ContactFormValueAdminController extends ApiController
 
     public function update(Request $request, $id)
     {
+        $contact_form_value = $this->contact_form_value_entity->find($id);
+        if (!$contact_form_value) {
+            throw new Exception('Contact form value not found');
+        }
         $data               = $request->all();
         $data['payload']    = json_encode($request->payload);
         $contact_form_value = $this->contact_form_value_repository->update($data, $id);
@@ -53,6 +62,10 @@ class ContactFormValueAdminController extends ApiController
 
     public function destroy($id)
     {
+        $contact_form_value = $this->contact_form_value_entity->find($id);
+        if (!$contact_form_value) {
+            throw new Exception('Contact form value not found');
+        }
         $this->contact_form_value_repository->destroy($id);
         return $this->success();
     }
